float support is deprecated for WaitForSingleObject
git-svn-id: http://svn.umitproject.org/svnroot/umit/trunk@4098 1105ee14-b0fa-0310-85a5-ff3eed429ff7

+++ umitCore/Scheduler.py
@@ -613,7 +613,7 @@ def run_scheduler(sched, winhndl=None):
             sleep_time = next_time - current_time + .1
             if winhndl:
                 stopsignal = win32event.WaitForSingleObject(winhndl,
-                        sleep_time * 1000)
+                        int(sleep_time * 1000))
                 if stopsignal == win32event.WAIT_OBJECT_0:
                     break
             else:
